[IMP] mail: refine fullscreen pulse animation
Purpose of this Commit:
The previous pulse animation for the full-screen button had an odd shape
and created an unnatural visual experience.

This commit updates the animation to a smooth circular pulse, resulting in a
more natural and soothing visual cue for the user when full-screen mode is
promoted.

task-5364914

closes odoo/odoo#237913

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>